backend/app/requirements.txt
aioapns==4.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiohttp-retry==2.9.1
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
APScheduler==3.11.0
asyncpg==0.30.0
attrs==25.3.0
bcrypt==4.3.0
boto3==1.35.96
botocore==1.35.99
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.2
deprecation==2.1.0
distro==1.9.0
dnspython==2.7.0
ecdsa==0.19.1
email-validator==2.1.0
fastapi==0.109.2
frozenlist==1.6.0
gotrue==2.12.0
gunicorn==23.0.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.27.2
hyperframe==6.1.0
idna==3.10
iniconfig==2.1.0
jiter==0.9.0
jmespath==1.0.1
multidict==6.4.3
openai==1.58.1
packaging==24.2
passlib==1.7.4
pillow==11.2.1
pluggy==1.6.0
postgrest==0.17.2
propcache==0.3.1
pyasn1==0.4.8
pycparser==2.22
pydantic==2.11.3
pydantic-settings==2.9.1
pydantic_core==2.33.1
Pygments==2.19.1
PyJWT==2.10.1
pyOpenSSL==25.1.0
pytest==8.4.0
pytest-mock==3.14.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.4.0
python-multipart==0.0.9
pytz==2025.2
realtime==2.5.0
requests==2.32.3
rsa==4.9.1
s3transfer==0.10.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.27
starlette==0.36.3
storage3==0.8.2
StrEnum==0.4.15
stripe==8.4.0
supabase==2.8.1
supafunc==0.6.2
tqdm==4.67.1
twilio==9.5.2
typing-inspection==0.4.0
typing_extensions==4.13.2
tzlocal==5.3.1
urllib3==1.26.20
uvicorn==0.27.1
websockets==12.0
yarl==1.20.0
